+++ server/tests/backends/sql_translator_unit_tests/test_google_big_query.py
@@ -1,4 +1,5 @@
 import pytest
+
 from weaverbird.backends.pypika_translator.translators.googlebigquery import (
     GoogleBigQueryTranslator,
 )