+++ server/tests/backends/sql_translator_unit_tests/test_type_mapping_translators.py
@@ -2,6 +2,7 @@ from typing import Any
 
 import pytest
 from pypika import Field, Query, functions
+
 from weaverbird.backends.pypika_translator.translators.base import DataTypeMapping, SQLTranslator
 from weaverbird.pipeline import steps