+++ server/tests/steps/test_fillna.py
@@ -1,10 +1,10 @@
 import numpy as np
 import pytest
 from pandas import DataFrame
-from weaverbird.backends.pandas_executor.steps.fillna import execute_fillna
-from weaverbird.pipeline.steps import FillnaStep
 
 from tests.utils import assert_dataframes_equals
+from weaverbird.backends.pandas_executor.steps.fillna import execute_fillna
+from weaverbird.pipeline.steps import FillnaStep
 
 
 @pytest.fixture