+++ server/tests/steps/test_text.py
@@ -5,10 +5,10 @@ from zoneinfo import ZoneInfo
 import numpy as np
 import pytest
 from pandas import DataFrame
-from weaverbird.backends.pandas_executor.steps.text import execute_text
-from weaverbird.pipeline.steps import TextStep
 
 from tests.utils import assert_dataframes_equals
+from weaverbird.backends.pandas_executor.steps.text import execute_text
+from weaverbird.pipeline.steps import TextStep
 
 
 @pytest.fixture