+++ server/tests/steps/test_uniquegroups.py
@@ -3,10 +3,10 @@ import random
 import numpy as np
 import pytest
 from pandas import DataFrame
-from weaverbird.backends.pandas_executor.steps.uniquegroups import execute_uniquegroups
-from weaverbird.pipeline.steps import UniqueGroupsStep
 
 from tests.utils import assert_dataframes_equals
+from weaverbird.backends.pandas_executor.steps.uniquegroups import execute_uniquegroups
+from weaverbird.pipeline.steps import UniqueGroupsStep
 
 
 @pytest.fixture